Add warn: max TX time 4ms with no ESB
Added warning per manufacturer guidance to limit transmissions not using
Enhanced Shock Burst features to 4ms bursts maximum or introduce delays
between sends.

The nRF24L01 stays in TX mode until it finishes transmitting a current
packet. If CE= 0 nRF24L01 returns
to standby-I mode. If CE= 1, the next action is determined by the status
of the TX FIFO. If the TX FIFO is
not empty the nRF24L01 remains in TX mode, transmitting the next packet.
If the TX FIFO is empty the
nRF24L01 goes into standby-II mode.The nRF24L01 transmitter PLL operates
in open loop when in TX
mode. It is important to never keep the nRF24L01 inTX mode for more than
4ms at a time. If the auto
retransmit is enabled, the nRF24L01 is never in TX mode long enough to
disobey this rule.